FY-227/fix(root): rfs.js->index.js
file이름 변경
